Added tests for realm repository

+++ src/repository/sqlite/claim_repository.js
@@ -21,18 +21,15 @@ export class ClaimRepositorySqlite implements ClaimRepository {
 
 
     constructor(theDBHelper: DBHelper, theRealmRepository: RealmRepository) {
+        if (!theDBHelper) {
+            throw new PersistenceError('db-helper not specified');
+        }
+        if (!theRealmRepository) {
+            throw new PersistenceError('realm-repository not specified');
+        }
         this.dbHelper           = theDBHelper;
         this.realmRepository    = theRealmRepository;
         this.sqlPrefix          = 'SELECT rowid AS id, realm_id, action, resource, condition FROM claims';
-    }
-
-    rowToClaim(row: any): Promise<Claim> {
-        return this.realmRepository.findById(row.realm_id).
-        then(realm => {
-            return new ClaimImpl(row.id, realm, row.action, row.resource, row.condition);
-        }).catch(err => {
-            throw new PersistenceError(`Could not find realm for claim ${row} due to ${err}`);
-        });
     }
 
     /**
@@ -40,6 +37,9 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * @param {*} id - database id
      */
     findById(id: number): Promise<Claim> {
+        if (!id) {
+            return Promise.reject(new PersistenceError('claim-id not specified'));
+        }
         var Claim = null;
         //
         return new Promise((resolve, reject) => {
@@ -47,7 +47,7 @@ export class ClaimRepositorySqlite implements ClaimRepository {
                 if (err) {
                     reject(new PersistenceError(`Could not find claim with id ${id} due to ${err}`));
                 } else if (row) {
-                    return this.rowToClaim(row).
+                    return this.__rowToClaim(row).
                     then(claim => {
                         resolve(claim);
                     }).catch(err => {
@@ -65,6 +65,9 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * @param {*} Claim - to save
      */
     save(claim: Claim): Promise<Claim> {
+        if (!claim) {
+            return Promise.reject(new PersistenceError('claim not specified'));
+        }
         let db = this.dbHelper.db;
         if (claim.id) {
             return new Promise((resolve, reject) => {
@@ -105,6 +108,9 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * @param {*} id - database id
      */
     removeById(id: number): Promise<boolean> {
+        if (!id) {
+            return Promise.reject(new PersistenceError('id not specified'));
+        }
         return new Promise((resolve, reject) => {
 			this.dbHelper.db.run('DELETE FROM claims WHERE rowid = ?', id, (err) => {
 				if (err) {
@@ -128,7 +134,7 @@ export class ClaimRepositorySqlite implements ClaimRepository {
     search(criteria: Map<string, any>, options?: QueryOptions): Promise<Array<Claim>> {
         let q:QueryHelper<Claim> = new QueryHelper(this.dbHelper.db);
         return q.query(this.sqlPrefix, criteria, (row) => {
-            return this.rowToClaim(row);
+            return this.__rowToClaim(row);
          }, options);
     }
 
@@ -136,6 +142,12 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * This method adds claims to principal
      */
     addClaimToPrincipal(principal: Principal, claim: Claim): Promise<void> {
+        if (!principal) {
+            return Promise.reject(new PersistenceError('principal not specified'));
+        }
+        if (!claim) {
+            return Promise.reject(new PersistenceError('claim not specified'));
+        }
         return new Promise((resolve, reject) => {
             this.dbHelper.db.run('INSERT INTO principals_claims VALUES (?, ?)', 
                 principal.id, claim.id, (err) => {
@@ -153,6 +165,12 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * This method adds claims to role
      */
     addClaimToRole(role: Role, claim: Claim): Promise<void> {
+        if (!role) {
+            return Promise.reject(new PersistenceError('role not specified'));
+        }
+        if (!claim) {
+            return Promise.reject(new PersistenceError('claim not specified'));
+        }
         return new Promise((resolve, reject) => {
             this.dbHelper.db.run('INSERT INTO roles_claims VALUES (?, ?)', 
                 role.id, claim.id, (err) => {
@@ -171,6 +189,12 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * This method removes claims from principal
      */
     removeClaimFromPrincipal(principal: Principal, claim: Claim): Promise<void> {
+        if (!principal) {
+            return Promise.reject(new PersistenceError('principal not specified'));
+        }
+        if (!claim) {
+            return Promise.reject(new PersistenceError('claim not specified'));
+        }
         return new Promise((resolve, reject) => {
             this.dbHelper.db.run('DELETE FROM principals_claims WHERE principal_id = ? and claim_id = ?', principal.id, claim.id, (err) => {
                 if (err) {
@@ -187,6 +211,12 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * This method remove claims from role
      */
     removeClaimFromRole(role: Role, claim: Claim) : Promise<void> {
+        if (!role) {
+            return Promise.reject(new PersistenceError('role not specified'));
+        }
+        if (!claim) {
+            return Promise.reject(new PersistenceError('claim not specified'));
+        }
         return new Promise((resolve, reject) => {
             this.dbHelper.db.run('DELETE FROM roles_claims WHERE role_id = ? and claim_id = ?', role.id, claim.id, (err) => {
                 if (err) {
@@ -203,6 +233,9 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * This method returns claims by principal that are associated to principal roles or directly to principal
      */
     loadPrincipalClaims(principal: Principal): Promise<void>  {
+        if (!principal) {
+            return Promise.reject(new PersistenceError('principal not specified'));
+        }
         let criteria: Map<string, any> = new Map();
         criteria.set('principal_id', principal.id);
 
@@ -211,7 +244,7 @@ export class ClaimRepositorySqlite implements ClaimRepository {
             'SELECT principal_id, claim_id AS id, realm_id, action, resource, condition ' +
             'FROM principals_claims INNER JOIN claims on claims.rowid = principals_claims.claim_id', 
             criteria, (row) => {
-            return this.rowToClaim(row).
+            return this.__rowToClaim(row).
             then(claim => {
                 principal.claims.add(claim);
                 return claim;
@@ -223,6 +256,9 @@ export class ClaimRepositorySqlite implements ClaimRepository {
      * This method load claims for role
      */
     loadRoleClaims(role: Role): Promise<void> {
+        if (!role) {
+            return Promise.reject(new PersistenceError('role not specified'));
+        }
         let criteria: Map<string, any> = new Map();
         criteria.set('role_id', role.id);
 
@@ -231,7 +267,7 @@ export class ClaimRepositorySqlite implements ClaimRepository {
             'SELECT role_id, claim_id AS id, realm_id, action, resource, condition ' +
             'FROM roles_claims INNER JOIN claims on claims.rowid = roles_claims.claim_id', 
             criteria, (row) => {
-            return this.rowToClaim(row).
+            return this.__rowToClaim(row).
             then(claim => {
                 role.claims.add(claim);
                 return claim;
@@ -239,5 +275,16 @@ export class ClaimRepositorySqlite implements ClaimRepository {
                 throw new PersistenceError(`Failed to load role claims for ${String(role)} due to ${err}`);
             })
          });
-    }    
+    }
+
+
+    __rowToClaim(row: any): Promise<Claim> {
+        return this.realmRepository.findById(row.realm_id).
+        then(realm => {
+            return new ClaimImpl(row.id, realm, row.action, row.resource, row.condition);
+        }).catch(err => {
+            throw new PersistenceError(`Could not find realm for claim ${row} due to ${err}`);
+        });
+    }
+
 }
