+++ src/repository/sqlite/principal_repository.js
@@ -24,6 +24,18 @@ export class PrincipalRepositorySqlite implements PrincipalRepository {
 
     constructor(theDBHelper: DBHelper, theRealmRepository: RealmRepository,
         theRoleRepository: RoleRepository, theClaimRepository: ClaimRepository) {
+        if (!theDBHelper) {
+            throw new PersistenceError('db-helper not specified');
+        }
+        if (!theRealmRepository) {
+            throw new PersistenceError('realm-repository not specified');
+        }
+        if (!theRoleRepository) {
+            throw new PersistenceError('role-repository not specified');
+        }
+        if (!theClaimRepository) {
+            throw new PersistenceError('claim-repository not specified');
+        }
         this.dbHelper           = theDBHelper;
         this.realmRepository    = theRealmRepository;
         this.roleRepository     = theRoleRepository;
@@ -31,30 +43,20 @@ export class PrincipalRepositorySqlite implements PrincipalRepository {
         this.sqlPrefix          = 'SELECT rowid AS id, principal_name FROM principals';
     }
 
-    rowToPrincipal(row: any): Promise<Principal> {
-        return this.realmRepository.findById(row.realm_id).
-            then(realm => {
-                let principal = new PrincipalImpl(row.id, realm, row.principal_name);
-                let promises = [this.claimRepository.loadPrincipalClaims(principal),
-                                this.roleRepository.loadPrincipalRoles(principal)];
-                return Promise.all(promises).
-                    then(result => {
-                    return principal;
-                });
-        });
-    }
-
     /**
      * This method finds principal by id
      * @param {*} id - database id
      */
     findById(id: number): Promise<Principal> {
+        if (!id) {
+            return Promise.reject(new PersistenceError('principal-id not specified'));
+        }
         return new Promise((resolve, reject) => {
             this.dbHelper.db.get(`${this.sqlPrefix} WHERE rowid == ?`, id, (err, row) => {
                 if (err) {
                     reject(new PersistenceError(`Could not find principal with id ${id}`));
                 } else if (row) {
-                    return this.rowToPrincipal(row).
+                    return this.__rowToPrincipal(row).
                         then(principal => {
                         resolve(principal);
                     });
@@ -70,12 +72,15 @@ export class PrincipalRepositorySqlite implements PrincipalRepository {
      * @param {*} principalName
      */
     findByName(principalName: string): Promise<Principal> {
+        if (!principalName) {
+            return Promise.reject(new PersistenceError('principalName not specified'));
+        }
         return new Promise((resolve, reject) => {
             this.dbHelper.db.get(`${this.sqlPrefix} WHERE principal_name == ?`, principalName, (err, row) => {
                 if (err) {
                     reject(new PersistenceError(`Could not find principal with name ${principalName}`));
                 } else if (row) {
-                    this.rowToPrincipal(row).
+                    this.__rowToPrincipal(row).
                         then(principal => {
                         resolve(principal);
                     });
@@ -91,6 +96,9 @@ export class PrincipalRepositorySqlite implements PrincipalRepository {
      * @param {*} principal - to save
      */
     save(principal: Principal): Promise<Principal> {
+        if (!principal) {
+            return Promise.reject(new PersistenceError('principal not specified'));
+        }
         if (principal.id) {
             throw new PersistenceError(`Principal is immutable and cannot be updated ${String(principal)}`);
         } else {
@@ -112,6 +120,9 @@ export class PrincipalRepositorySqlite implements PrincipalRepository {
      * @param {*} id - database id
      */
     removeById(id: number): Promise<boolean> {
+        if (!id) {
+            return Promise.reject(new PersistenceError('id not specified'));
+        }
         return new Promise((resolve, reject) => {
 			this.dbHelper.db.run('DELETE FROM principals WHERE rowid = ?', id, (err) => {
 				if (err) {
@@ -131,7 +142,21 @@ export class PrincipalRepositorySqlite implements PrincipalRepository {
     search(criteria: Map<string, any>, options?: QueryOptions): Promise<Array<Principal>> {
         let q:QueryHelper<Principal> = new QueryHelper(this.dbHelper.db);
         return q.query(this.sqlPrefix, criteria, (row) => {
-             return this.rowToPrincipal(row);
+             return this.__rowToPrincipal(row);
          }, options);
     }
+
+    __rowToPrincipal(row: any): Promise<Principal> {
+        return this.realmRepository.findById(row.realm_id).
+            then(realm => {
+                let principal = new PrincipalImpl(row.id, realm, row.principal_name);
+                let promises = [this.claimRepository.loadPrincipalClaims(principal),
+                                this.roleRepository.loadPrincipalRoles(principal)];
+                return Promise.all(promises).
+                    then(result => {
+                    return principal;
+                });
+        });
+    }
+
 }
